+++ deepchem/trans/tests/test_transformers.py
@@ -2,12 +2,7 @@
 Tests for transformer objects.
 """
 from deepchem.molnet import load_delaney
-from deepchem.trans.transformers import FeaturizationTransformer
 from deepchem.trans.transformers import DataTransforms
-
-__author__ = "Bharath Ramsundar"
-__copyright__ = "Copyright 2016, Stanford University"
-__license__ = "MIT"
 
 import os
 import unittest
@@ -31,44 +26,6 @@ def load_solubility_data():
   return loader.create_dataset(input_file)
 
 
-def load_feat_multitask_data():
-  """Load example with numerical features, tasks."""
-  current_dir = os.path.dirname(os.path.abspath(__file__))
-  features = ["feat0", "feat1", "feat2", "feat3", "feat4", "feat5"]
-  featurizer = dc.feat.UserDefinedFeaturizer(features)
-  tasks = ["task0", "task1", "task2", "task3", "task4", "task5"]
-  input_file = os.path.join(current_dir,
-                            "../../models/tests/feat_multitask_example.csv")
-  loader = dc.data.UserCSVLoader(
-      tasks=tasks, featurizer=featurizer, id_field="id")
-  return loader.featurize(input_file)
-
-
-def load_gaussian_cdf_data():
-  """Load example with numbers sampled from Gaussian normal distribution.
-     Each feature and task is a column of values that is sampled
-     from a normal distribution of mean 0, stdev 1."""
-  current_dir = os.path.dirname(os.path.abspath(__file__))
-  features = ["feat0", "feat1"]
-  featurizer = dc.feat.UserDefinedFeaturizer(features)
-  tasks = ["task0", "task1"]
-  input_file = os.path.join(current_dir,
-                            "../../models/tests/gaussian_cdf_example.csv")
-  loader = dc.data.UserCSVLoader(
-      tasks=tasks, featurizer=featurizer, id_field="id")
-  return loader.featurize(input_file)
-
-
-def load_unlabelled_data():
-  current_dir = os.path.dirname(os.path.abspath(__file__))
-  featurizer = dc.feat.CircularFingerprint(size=1024)
-  tasks = []
-  input_file = os.path.join(current_dir, "../../data/tests/no_labels.csv")
-  loader = dc.data.CSVLoader(
-      tasks=tasks, smiles_field="smiles", featurizer=featurizer)
-  return loader.featurize(input_file)
-
-
 class TestTransformers(unittest.TestCase):
   """
   Test top-level API for transformer objects.
@@ -87,248 +44,6 @@ class TestTransformers(unittest.TestCase):
     # reshaping the vector to image
     data = np.reshape(data, (28, 28))
     self.d = data
-
-  def test_y_log_transformer(self):
-    """Tests logarithmic data transformer."""
-    solubility_dataset = load_solubility_data()
-    log_transformer = dc.trans.LogTransformer(
-        transform_y=True, dataset=solubility_dataset)
-    X, y, w, ids = (solubility_dataset.X, solubility_dataset.y,
-                    solubility_dataset.w, solubility_dataset.ids)
-    solubility_dataset = log_transformer.transform(solubility_dataset)
-    X_t, y_t, w_t, ids_t = (solubility_dataset.X, solubility_dataset.y,
-                            solubility_dataset.w, solubility_dataset.ids)
-
-    # Check ids are unchanged.
-    for id_elt, id_t_elt in zip(ids, ids_t):
-      assert id_elt == id_t_elt
-    # Check X is unchanged since this is a y transformer
-    np.testing.assert_allclose(X, X_t)
-    # Check w is unchanged since this is a y transformer
-    np.testing.assert_allclose(w, w_t)
-    # Check y is now a logarithmic version of itself
-    np.testing.assert_allclose(y_t, np.log(y + 1))
-
-    # Check that untransform does the right thing.
-    np.testing.assert_allclose(log_transformer.untransform(y_t), y)
-
-  def test_transform_unlabelled(self):
-    ul_dataset = load_unlabelled_data()
-    # transforming y should raise an exception
-    with self.assertRaises(ValueError) as context:
-      dc.trans.transformers.Transformer(transform_y=True).transform(ul_dataset)
-
-    # transforming w should raise an exception
-    with self.assertRaises(ValueError) as context:
-      dc.trans.transformers.Transformer(transform_w=True).transform(ul_dataset)
-
-    # transforming X should be okay
-    dc.trans.NormalizationTransformer(
-        transform_X=True, dataset=ul_dataset).transform(ul_dataset)
-
-  def test_X_log_transformer(self):
-    """Tests logarithmic data transformer."""
-    solubility_dataset = load_solubility_data()
-    log_transformer = dc.trans.LogTransformer(
-        transform_X=True, dataset=solubility_dataset)
-    X, y, w, ids = (solubility_dataset.X, solubility_dataset.y,
-                    solubility_dataset.w, solubility_dataset.ids)
-    solubility_dataset = log_transformer.transform(solubility_dataset)
-    X_t, y_t, w_t, ids_t = (solubility_dataset.X, solubility_dataset.y,
-                            solubility_dataset.w, solubility_dataset.ids)
-
-    # Check ids are unchanged.
-    for id_elt, id_t_elt in zip(ids, ids_t):
-      assert id_elt == id_t_elt
-    # Check y is unchanged since this is a X transformer
-    np.testing.assert_allclose(y, y_t)
-    # Check w is unchanged since this is a y transformer
-    np.testing.assert_allclose(w, w_t)
-    # Check y is now a logarithmic version of itself
-    np.testing.assert_allclose(X_t, np.log(X + 1))
-
-    # Check that untransform does the right thing.
-    np.testing.assert_allclose(log_transformer.untransform(X_t), X)
-
-  def test_y_log_transformer_select(self):
-    """Tests logarithmic data transformer with selection."""
-    multitask_dataset = load_feat_multitask_data()
-    dfe = pd.read_csv(
-        os.path.join(self.current_dir,
-                     "../../models/tests/feat_multitask_example.csv"))
-    tid = []
-    tasklist = ["task0", "task3", "task4", "task5"]
-    first_task = "task0"
-    for task in tasklist:
-      tiid = dfe.columns.get_loc(task) - dfe.columns.get_loc(first_task)
-      tid = np.concatenate((tid, np.array([tiid])))
-    tasks = tid.astype(int)
-    log_transformer = dc.trans.LogTransformer(
-        transform_y=True, tasks=tasks, dataset=multitask_dataset)
-    X, y, w, ids = (multitask_dataset.X, multitask_dataset.y,
-                    multitask_dataset.w, multitask_dataset.ids)
-    multitask_dataset = log_transformer.transform(multitask_dataset)
-    X_t, y_t, w_t, ids_t = (multitask_dataset.X, multitask_dataset.y,
-                            multitask_dataset.w, multitask_dataset.ids)
-
-    # Check ids are unchanged.
-    for id_elt, id_t_elt in zip(ids, ids_t):
-      assert id_elt == id_t_elt
-    # Check X is unchanged since this is a y transformer
-    np.testing.assert_allclose(X, X_t)
-    # Check w is unchanged since this is a y transformer
-    np.testing.assert_allclose(w, w_t)
-    # Check y is now a logarithmic version of itself
-    np.testing.assert_allclose(y_t[:, tasks], np.log(y[:, tasks] + 1))
-
-    # Check that untransform does the right thing.
-    np.testing.assert_allclose(log_transformer.untransform(y_t), y)
-
-  def test_X_log_transformer_select(self):
-    # Tests logarithmic data transformer with selection.
-    multitask_dataset = load_feat_multitask_data()
-    dfe = pd.read_csv(
-        os.path.join(self.current_dir,
-                     "../../models/tests/feat_multitask_example.csv"))
-    fid = []
-    featurelist = ["feat0", "feat1", "feat2", "feat3", "feat5"]
-    first_feature = "feat0"
-    for feature in featurelist:
-      fiid = dfe.columns.get_loc(feature) - dfe.columns.get_loc(first_feature)
-      fid = np.concatenate((fid, np.array([fiid])))
-    features = fid.astype(int)
-    log_transformer = dc.trans.LogTransformer(
-        transform_X=True, features=features, dataset=multitask_dataset)
-    X, y, w, ids = (multitask_dataset.X, multitask_dataset.y,
-                    multitask_dataset.w, multitask_dataset.ids)
-    multitask_dataset = log_transformer.transform(multitask_dataset)
-    X_t, y_t, w_t, ids_t = (multitask_dataset.X, multitask_dataset.y,
-                            multitask_dataset.w, multitask_dataset.ids)
-
-    # Check ids are unchanged.
-    for id_elt, id_t_elt in zip(ids, ids_t):
-      assert id_elt == id_t_elt
-    # Check y is unchanged since this is a X transformer
-    np.testing.assert_allclose(y, y_t)
-    # Check w is unchanged since this is a y transformer
-    np.testing.assert_allclose(w, w_t)
-    # Check y is now a logarithmic version of itself
-    np.testing.assert_allclose(X_t[:, features], np.log(X[:, features] + 1))
-
-    # Check that untransform does the right thing.
-    np.testing.assert_allclose(log_transformer.untransform(X_t), X)
-
-  def test_y_normalization_transformer(self):
-    """Tests normalization transformer."""
-    solubility_dataset = load_solubility_data()
-    normalization_transformer = dc.trans.NormalizationTransformer(
-        transform_y=True, dataset=solubility_dataset)
-    X, y, w, ids = (solubility_dataset.X, solubility_dataset.y,
-                    solubility_dataset.w, solubility_dataset.ids)
-    solubility_dataset = normalization_transformer.transform(solubility_dataset)
-    X_t, y_t, w_t, ids_t = (solubility_dataset.X, solubility_dataset.y,
-                            solubility_dataset.w, solubility_dataset.ids)
-    # Check ids are unchanged.
-    for id_elt, id_t_elt in zip(ids, ids_t):
-      assert id_elt == id_t_elt
-    # Check X is unchanged since this is a y transformer
-    np.testing.assert_allclose(X, X_t)
-    # Check w is unchanged since this is a y transformer
-    np.testing.assert_allclose(w, w_t)
-    # Check that y_t has zero mean, unit std.
-    assert np.isclose(y_t.mean(), 0.)
-    assert np.isclose(y_t.std(), 1.)
-
-    # Check that untransform does the right thing.
-    np.testing.assert_allclose(normalization_transformer.untransform(y_t), y)
-
-  def test_X_normalization_transformer(self):
-    """Tests normalization transformer."""
-    solubility_dataset = load_solubility_data()
-    normalization_transformer = dc.trans.NormalizationTransformer(
-        transform_X=True, dataset=solubility_dataset)
-    X, y, w, ids = (solubility_dataset.X, solubility_dataset.y,
-                    solubility_dataset.w, solubility_dataset.ids)
-    solubility_dataset = normalization_transformer.transform(solubility_dataset)
-    X_t, y_t, w_t, ids_t = (solubility_dataset.X, solubility_dataset.y,
-                            solubility_dataset.w, solubility_dataset.ids)
-    # Check ids are unchanged.
-    for id_elt, id_t_elt in zip(ids, ids_t):
-      assert id_elt == id_t_elt
-    # Check y is unchanged since this is a X transformer
-    np.testing.assert_allclose(y, y_t)
-    # Check w is unchanged since this is a y transformer
-    np.testing.assert_allclose(w, w_t)
-    # Check that X_t has zero mean, unit std.
-    # np.set_printoptions(threshold='nan')
-    mean = X_t.mean(axis=0)
-    assert np.amax(np.abs(mean - np.zeros_like(mean))) < 1e-7
-    orig_std_array = X.std(axis=0)
-    std_array = X_t.std(axis=0)
-    # Entries with zero std are not normalized
-    for orig_std, std in zip(orig_std_array, std_array):
-      if not np.isclose(orig_std, 0):
-        assert np.isclose(std, 1)
-
-    # TODO(rbharath): Untransform doesn't work properly for binary feature
-    # vectors. Need to figure out what's wrong here. (low priority)
-    ## Check that untransform does the right thing.
-    # np.testing.assert_allclose(normalization_transformer.untransform(X_t), X)
-
-  def test_cdf_X_transformer(self):
-    """Test CDF transformer on Gaussian normal dataset."""
-    target = np.array(np.transpose(np.linspace(0., 1., 1001)))
-    target = np.transpose(np.array(np.append([target], [target], axis=0)))
-    gaussian_dataset = load_gaussian_cdf_data()
-    bins = 1001
-    cdf_transformer = dc.trans.CDFTransformer(
-        transform_X=True, dataset=gaussian_dataset, bins=bins)
-    X, y, w, ids = (gaussian_dataset.X, gaussian_dataset.y, gaussian_dataset.w,
-                    gaussian_dataset.ids)
-    gaussian_dataset = cdf_transformer.transform(gaussian_dataset, bins=bins)
-    X_t, y_t, w_t, ids_t = (gaussian_dataset.X, gaussian_dataset.y,
-                            gaussian_dataset.w, gaussian_dataset.ids)
-
-    # Check ids are unchanged.
-    for id_elt, id_t_elt in zip(ids, ids_t):
-      assert id_elt == id_t_elt
-    # Check y is unchanged since this is an X transformer
-    np.testing.assert_allclose(y, y_t)
-    # Check w is unchanged since this is an X transformer
-    np.testing.assert_allclose(w, w_t)
-    # Check X is now holding the proper values when sorted.
-    sorted = np.sort(X_t, axis=0)
-    np.testing.assert_allclose(sorted, target)
-
-  def test_cdf_y_transformer(self):
-    # Test CDF transformer on Gaussian normal dataset.
-    target = np.array(np.transpose(np.linspace(0., 1., 1001)))
-    target = np.transpose(np.array(np.append([target], [target], axis=0)))
-    gaussian_dataset = load_gaussian_cdf_data()
-    bins = 1001
-    cdf_transformer = dc.trans.CDFTransformer(
-        transform_y=True, dataset=gaussian_dataset, bins=bins)
-    X, y, w, ids = (gaussian_dataset.X, gaussian_dataset.y, gaussian_dataset.w,
-                    gaussian_dataset.ids)
-    gaussian_dataset = cdf_transformer.transform(gaussian_dataset, bins=bins)
-    X_t, y_t, w_t, ids_t = (gaussian_dataset.X, gaussian_dataset.y,
-                            gaussian_dataset.w, gaussian_dataset.ids)
-
-    # Check ids are unchanged.
-    for id_elt, id_t_elt in zip(ids, ids_t):
-      assert id_elt == id_t_elt
-    # Check X is unchanged since this is an y transformer
-    np.testing.assert_allclose(X, X_t)
-    # Check w is unchanged since this is an y transformer
-    np.testing.assert_allclose(w, w_t)
-    # Check y is now holding the proper values when sorted.
-    sorted = np.sort(y_t, axis=0)
-    np.testing.assert_allclose(sorted, target)
-
-    # Check that untransform does the right thing.
-    y_restored = cdf_transformer.untransform(y_t)
-    assert np.max(y_restored - y) < 1e-5
-    #np.testing.assert_allclose(y_restored, y)
 
   def test_clipping_X_transformer(self):
     """Test clipping transformer on X of singletask dataset."""
@@ -382,59 +97,6 @@ class TestTransformers(unittest.TestCase):
     # Check y is now holding the proper values when sorted.
     np.testing.assert_allclose(y_t, target)
 
-  def test_power_X_transformer(self):
-    """Test Power transformer on Gaussian normal dataset."""
-    gaussian_dataset = load_gaussian_cdf_data()
-    powers = [1, 2, 0.5]
-    power_transformer = dc.trans.PowerTransformer(
-        transform_X=True, powers=powers)
-    X, y, w, ids = (gaussian_dataset.X, gaussian_dataset.y, gaussian_dataset.w,
-                    gaussian_dataset.ids)
-    gaussian_dataset2 = power_transformer.transform(gaussian_dataset)
-    X_t, y_t, w_t, ids_t = (gaussian_dataset2.X, gaussian_dataset2.y,
-                            gaussian_dataset2.w, gaussian_dataset2.ids)
-
-    # Check ids are unchanged.
-    for id_elt, id_t_elt in zip(ids, ids_t):
-      assert id_elt == id_t_elt
-    # Check y is unchanged since this is an X transformer
-    np.testing.assert_allclose(y, y_t)
-    # Check w is unchanged since this is an X transformer
-    np.testing.assert_allclose(w, w_t)
-    # Check X is now holding the proper values in each column.
-    np.testing.assert_allclose(X_t.shape[1], len(powers) * X.shape[1])
-    np.testing.assert_allclose(X, X_t[:, :2])
-    np.testing.assert_allclose(np.power(X, 2), X_t[:, 2:4])
-    np.testing.assert_allclose(np.power(X, 0.5), X_t[:, 4:])
-
-  def test_power_y_transformer(self):
-    """Test Power transformer on Gaussian normal dataset."""
-    gaussian_dataset = load_gaussian_cdf_data()
-    powers = [1, 2, 0.5]
-    power_transformer = dc.trans.PowerTransformer(
-        transform_y=True, powers=powers)
-    X, y, w, ids = (gaussian_dataset.X, gaussian_dataset.y, gaussian_dataset.w,
-                    gaussian_dataset.ids)
-    gaussian_dataset2 = power_transformer.transform(gaussian_dataset)
-    X_t, y_t, w_t, ids_t = (gaussian_dataset2.X, gaussian_dataset2.y,
-                            gaussian_dataset2.w, gaussian_dataset2.ids)
-
-    # Check ids are unchanged.
-    for id_elt, id_t_elt in zip(ids, ids_t):
-      assert id_elt == id_t_elt
-    # Check X is unchanged since this is an X transformer
-    np.testing.assert_allclose(X, X_t)
-    # Check w is unchanged since this is an X transformer
-    np.testing.assert_allclose(w, w_t)
-    # Check y is now holding the proper values in each column.
-    np.testing.assert_allclose(y_t.shape[1], len(powers) * y.shape[1])
-    np.testing.assert_allclose(y, y_t[:, :2])
-    np.testing.assert_allclose(np.power(y, 2), y_t[:, 2:4])
-    np.testing.assert_allclose(np.power(y, 0.5), y_t[:, 4:])
-
-    # Check that untransform does the right thing.
-    np.testing.assert_allclose(power_transformer.untransform(y_t), y)
-
   def test_coulomb_fit_transformer(self):
     """Test coulomb fit transformer on singletask dataset."""
     n_samples = 10
@@ -473,19 +135,6 @@ class TestTransformers(unittest.TestCase):
     assert np.allclose(test_dataset_trans.X[0, :10], sims[:10])
     assert np.allclose(test_dataset_trans.X[0, 10:20], [0] * 10)
     assert not np.isclose(dataset_trans.X[0, 0], 1.)
-
-  def test_featurization_transformer(self):
-    fp_size = 2048
-    tasks, all_dataset, transformers = load_delaney('Raw')
-    train = all_dataset[0]
-    transformer = FeaturizationTransformer(
-        transform_X=True,
-        dataset=train,
-        featurizer=dc.feat.CircularFingerprint(size=fp_size))
-    new_train = transformer.transform(train)
-
-    self.assertEqual(new_train.y.shape, train.y.shape)
-    self.assertEqual(new_train.X.shape[-1], fp_size)
 
   def test_blurring(self):
     # Check Blurring
@@ -592,27 +241,6 @@ class TestTransformers(unittest.TestCase):
     np.random.seed(0)
     check_random_noise = dt.salt_pepper_noise(prob, salt=255, pepper=0)
     assert np.allclose(random_noise, check_random_noise)
-
-  def test_DAG_transformer(self):
-    """Tests the DAG transformer."""
-    np.random.seed(123)
-    tf.random.set_seed(123)
-    n_tasks = 1
-
-    # Load mini log-solubility dataset.
-    featurizer = dc.feat.ConvMolFeaturizer()
-    tasks = ["outcome"]
-    input_file = os.path.join(self.current_dir,
-                              "../../models/tests/example_regression.csv")
-    loader = dc.data.CSVLoader(
-        tasks=tasks, smiles_field="smiles", featurizer=featurizer)
-    dataset = loader.create_dataset(input_file)
-    transformer = dc.trans.DAGTransformer(max_atoms=50)
-    dataset = transformer.transform(dataset)
-    # The transformer generates n DAGs for a molecule with n
-    # atoms. These are denoted the "parents"
-    for idm, mol in enumerate(dataset.X):
-      assert dataset.X[idm].get_num_atoms() == len(dataset.X[idm].parents)
 
   def test_median_filter(self):
     #Check median filter
